refactor: improve SpeakTouchService dependencies && packages
