+++ app/src/main/java/com/neo/speaktouch/intercepter/event/CallbackInterceptor.kt
@@ -16,11 +16,12 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.neo.speaktouch.intercepter.gesture
+package com.neo.speaktouch.intercepter.event
 
 import android.view.accessibility.AccessibilityEvent
 import android.view.accessibility.AccessibilityNodeInfo
-import com.neo.speaktouch.intercepter.event.EventInterceptor
+import com.neo.speaktouch.intercepter.event.contract.EventInterceptor
+import dagger.hilt.android.scopes.ServiceScoped
 import javax.inject.Inject
 
 @ServiceScoped
@@ -47,8 +48,7 @@ class CallbackInterceptor @Inject constructor() : EventInterceptor {
         scrolls.add(callback)
     }
 
-}
-
-abstract class Scroll(val target: AccessibilityNodeInfo) {
-    abstract operator fun invoke()
+    abstract class Scroll(val target: AccessibilityNodeInfo) {
+        abstract operator fun invoke()
+    }
 }
