+++ app/src/main/java/com/neo/speaktouch/intercepter/event/FocusInterceptor.kt
@@ -20,7 +20,8 @@ package com.neo.speaktouch.intercepter.event
 
 import android.view.accessibility.AccessibilityEvent
 import androidx.core.view.accessibility.AccessibilityNodeInfoCompat
-import com.neo.speaktouch.utils.`object`.NodeValidator
+import com.neo.speaktouch.intercepter.event.contract.EventInterceptor
+import com.neo.speaktouch.utils.NodeValidator
 import com.neo.speaktouch.utils.extension.getNearestAncestor
 import dagger.hilt.android.scopes.ServiceScoped
 import javax.inject.Inject
